+++ Actions/TelemetryHelper.psm1
@@ -15,7 +15,7 @@ $signals = @{
     "DO0083" = "AL-Go action ran: BuildPowerPlatform"
 
     "DO0090" = "AL-Go workflow ran: AddExistingAppOrTestApp"
-    "DO0091" = "AL-Go workflow ran: CiCd"
+    "DO0091" = "AL-Go workflow ran: CICD"
     "DO0092" = "AL-Go workflow ran: CreateApp"
     "DO0093" = "AL-Go workflow ran: CreateOnlineDevelopmentEnvironment"
     "DO0094" = "AL-Go workflow ran: CreateRelease"
@@ -28,6 +28,7 @@ $signals = @{
     "DO0101" = "AL-Go workflow ran: Current"
     "DO0102" = "AL-Go workflow ran: CreatePerformanceTestApp"
     "DO0103" = "AL-Go workflow ran: PublishToAppSource"
+    "DO0104" = "AL-Go workflow ran: PullRequestHandler"
 }
 ## Andersg <-- add your actions and worksflows above
 Function strToHexStr {