+++ e2eTests/Test-AL-Go.ps1
@@ -128,11 +128,11 @@ $runs++
 $run = Run-CICD -wait -branch $branch
 $runs++
 
-Test-NumberOfRuns -expectedNumberOfRuns $runs
+Test-NumberOfRuns -expectedNumberOfRuns $runs -repository $repository
 Test-ArtifactsFromRun -runid $run.id -expectedArtifacts @{"Apps"=2;"TestApps"=1} -expectedNumberOfTests 1 -folder 'artifacts' -repoVersion '1.0' -appVersion ''
 
 # Create Release
-Run-CreateRelease -appVersion "1.0.$($runs-2).0" -name '1.0' -tag '1.0' -wait -branch $branch | Out-Null
+Run-CreateRelease -appVersion "1.0.$($runs-2).0" -name 'v1.0' -tag '1.0.0' -wait -branch $branch | Out-Null
 $runs++
 
 # Create New App
@@ -186,7 +186,7 @@ if ($multiProject) {
 else {
     Test-ArtifactsFromRun -runid $run.id -expectedArtifacts @{"Apps"=3;"TestApps"=2} -expectedNumberOfTests 2 -folder 'artifacts2' -repoVersion '2.0' -appVersion ''
 }
-Test-NumberOfRuns -expectedNumberOfRuns $runs
+Test-NumberOfRuns -expectedNumberOfRuns $runs -repository $repository
 
 # Modify versioning strategy
 $projectSettingsFiles | Select-Object -Last 1 | ForEach-Object {
@@ -220,7 +220,7 @@ if (!(Test-Path "$($project1Folder).AL-Go\*.ps1")) { throw "Local PowerShell scr
 if (!(Test-Path ".github\workflows\AddExistingAppOrTestApp.yaml")) { throw "AddExistingAppOrTestApp.yaml was not updated by Update AL-Go System Files" }
 
 # Create Release
-Run-CreateRelease -appVersion latest -name "v3.0" -tag "v3.0" -wait -branch $branch | Out-Null
+Run-CreateRelease -appVersion latest -name "v3.0" -tag "3.0.0" -wait -branch $branch | Out-Null
 $runs++
 
 # Launch Current, NextMinor and NextMajor builds
@@ -247,7 +247,7 @@ $runs++
 WaitWorkflow -runid $runTestCurrent.id -noDelay
 $runs++
 
-Test-NumberOfRuns -expectedNumberOfRuns $runs
+Test-NumberOfRuns -expectedNumberOfRuns $runs -repository $repository
 
 Set-Location $prevLocation